pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataset-validator"
version = "0.1.0"
description = "A framework to validate and sanitize datasets for sensitivity and compliance"
authors = [
    { name = "Chandrakanth R M", email = "chandrakanthrm89@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "numpy",
    "google-generativeai",
    "pandas"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8"
]
